refactor!: move utils to utils (#389)
* docs: adds utils to styleguide

* refactor(touchcapture)!: move to utils directory

* refactor(transitionfadein)!: move to utils directory

* refactor(transitionresize)!: move to utils directory

* refactor(transitionspringleft)!: move to utils directory

* refactor(transitionspringup)!: move to utils directory

* refactor(transitionstaggered)!: move to utils directory
