+++ src/utils/TransitionSpringUp/README.md
@@ -1,23 +1,23 @@
-# Transition Fade In
+# Transition Spring Up
 
 ```vue
 <template>
 	<div class="container">
-		<m-transition-fade-in>
+		<m-transition-spring-up>
 			<div
 				v-if="visible"
 				class="gray-box"
 			/>
-		</m-transition-fade-in>
+		</m-transition-spring-up>
 	</div>
 </template>
 
 <script>
-import { MTransitionFadeIn } from '@square/maker/components/TransitionFadeIn';
+import { MTransitionSpringUp } from '@square/maker/utils/TransitionSpringUp';
 
 export default {
 	components: {
-		MTransitionFadeIn,
+		MTransitionSpringUp,
 	},
 
 	data() {
@@ -43,8 +43,9 @@ export default {
 }
 
 .gray-box {
-    width: 300px;
-    height: 500px;
+    width: 100%;
+    height: 100%;
+    border-radius: 16px 16px 0 0;
     background-color: #ccc;
 }
 </style>
@@ -53,7 +54,7 @@ export default {
 <!-- api-tables:start -->
 ## Slots
 
-| Slot    | Description              |
-| ------- | ------------------------ |
-| default | content to fade in & out |
+| Slot    | Description          |
+| ------- | -------------------- |
+| default | content to spring up |
 <!-- api-tables:end -->
